requirement.txt
appdirs==1.4.4
apturl==0.5.2
arrow==0.15.8
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.4
bcrypt==3.1.7
beautifulsoup4==4.9.1
binaryornot==0.4.4
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2020.6.20
chardet==3.0.4
cliapp==1.20180812.1
click==7.1.2
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
cookiecutter==1.7.2
countries==0.2.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.0.9
django-admin-sortable2==0.7.7
django-angular==2.3
django-appconf==1.0.4
django-bootstrap-modal-forms==1.5.0
django-bootstrap3==12.1.0
django-bootstrap4==2.2.0
django-classy-tags==1.0.0
django-cms==3.7.4
django-constance==2.8.0
django-countries==6.1.3
django-crispy-forms==1.8.1
django-dynamic-form-fields==0.3.2
django-entangled==0.3
django-filer==1.7.1
django-filter==2.2.0
django-formtools==2.2
django-fsm==2.7.0
django-fsm-admin==1.2.4
django-ipware==3.0.0
django-js-asset==1.2.2
django-jsonify==0.3.0
django-money==1.1
django-mptt==0.11.0
django-polymorphic==2.1.2
django-popup-view-field==0.6.1
django-post-office==3.4.1
django-rest-auth==0.9.5
django-sekizai==1.1.0
django-select2==7.4.2
django-shop==1.2.3
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangocms-admin-style==1.5.0
djangocms-attributes-field==1.2.0
djangocms-bootstrap4==1.6.0
djangocms-cascade==1.3.3
djangocms-file==2.4.0
djangocms-googlemap==1.4.0
djangocms-icon==1.5.0
djangocms-installer==1.2.3
djangocms-link==2.6.1
djangocms-picture==2.4.0
djangocms-snippet==2.3.0
djangocms-style==2.3.0
djangocms-text-ckeditor==3.10.0
djangocms-video==2.3.0
djangorestframework==3.11.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
easy-thumbnails==2.7
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
html5lib==1.1
httplib2==0.14.0
idna==2.10
importlib-metadata==1.5.0
isort==5.7.0
Jinja2==2.11.2
jinja2-time==0.2.0
jsonfield==3.1.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
mysql-connector==2.2.9
mysql-connector-python==8.0.15
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
pbr==5.5.1
pexpect==4.6.0
Pillow==7.0.0
pipenv==2020.6.2
poyo==0.5.0
protobuf==3.6.1
py-moneyed==0.8.0
pycairo==1.16.2
pycodestyle==2.6.0
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.15.5
PySimpleGUI==4.18.0
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-slugify==4.0.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
redis==3.5.3
reportlab==3.5.34
requests==2.24.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
soupsieve==2.0.1
South==1.0.2
sqlparse==0.3.0
ssh-import-id==5.10
sweetify==2.0.0
systemd-python==234
testresources==2.0.1
text-unidecode==1.3
texttable==1.6.2
toml==0.10.1
ttystatus==0.38
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.1
urllib3==1.25.10
usb-creator==0.3.7
virtualenv==20.0.30
virtualenv-clone==0.5.4
wadllib==1.3.3
webencodings==0.5.1
websocket-client==0.53.0
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
